requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
aiohttp==3.10.5
aiosqlite==0.20.0
jinja2==3.1.4
httpx==0.27.2
python-dotenv==1.0.1
